+++ src/main/kotlin/ru/meanmail/RequirementType.kt
@@ -4,7 +4,6 @@ enum class RequirementType(val kind: String) {
     NAME("package"),
     URL("package"),
     REFERENCE("file"),
-    CONSTRAINT("file"),
     EDITABLE("url"),
     PATH("path")
 }