+++ src/main/kotlin/ru/meanmail/psi/ConstraintReq.kt
@@ -1,21 +1,7 @@
 package ru.meanmail.psi
 
-import ru.meanmail.RequirementType
-
-interface ConstraintReq : NamedElement, Requirement {
+interface ConstraintReq : NamedElement {
 
     val uriReference: UriReference?
-
-    override fun enabled(values: Map<String, String?>): Boolean {
-        return true
-    }
-
-    override val displayName: String
-        get() {
-            return uriReference?.text ?: ""
-        }
-
-    override val type: RequirementType
-        get() = RequirementType.CONSTRAINT
 
 }