+++ src/main/kotlin/ru/meanmail/psi/NameReq.kt
@@ -23,11 +23,6 @@ interface NameReq : NamedElement, Requirement {
             return name.text
         }
 
-    val requirement: String
-        get() {
-            return "${name.text}${extras?.text ?: ""}${versionspec?.text ?: ""}"
-        }
-
     override val type: RequirementType
         get() = RequirementType.NAME